# docs/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "Body_import_outages_api_v1_outages_import_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_outages_api_v1_outages_import_post",
        "type": "object"
      },
      "BulkOutageCreate": {
        "example": {
          "outages": [
            {
              "affected_services": [
                "core-api"
              ],
              "assigned_to": "oncall@example.com",
              "created_by": "reporter@example.com",
              "description": "Example outage affecting core API.",
              "detected_at": "2026-01-01T00:00:00Z",
              "id": "outage-001",
              "location": {
                "latitude": 40.7128,
                "longitude": -74.006
              },
              "severity": "high",
              "site_id": "site-A",
              "site_name": "Site A",
              "status": "open"
            }
          ]
        },
        "properties": {
          "outages": {
            "items": {
              "$ref": "#/components/schemas/OutageCreate"
            },
            "title": "Outages",
            "type": "array"
          }
        },
        "required": [
          "outages"
        ],
        "title": "BulkOutageCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportConsistency": {
        "enum": [
          "atomic",
          "partial"
        ],
        "title": "ImportConsistency",
        "type": "string"
      },
...
      "OutageStatus": {
        "enum": [
          "open",
          "resolved"
        ],
        "title": "OutageStatus",
        "type": "string"
      },
...
      "Severity": {
        "enum": [
          "critical",
          "high",
          "medium",
          "low"
        ],
        "title": "Severity",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "ApexChainx Backend API",
    "title": "ApexChainx API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/outages/bulk": {
      "post": {
        "operationId": "bulk_create_outages_api_v1_outages_bulk_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkOutageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Bulk Create Outages Api V1 Outages Bulk Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Create Outages",
        "tags": [
          "outages"
        ]
      }
    },
    "/api/v1/outages/export": {
      "get": {
        "operationId": "export_outages_endpoint_api_v1_outages_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Severity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OutageStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Outages Endpoint",
        "tags": [
          "outages"
        ]
      }
    },
    "/api/v1/outages/import": {
      "post": {
        "operationId": "import_outages_api_v1_outages_import_post",
        "parameters": [
          {
            "description": "Validation-only mode: validate all rows WITHOUT persisting to database. Returns the same field/row-level errors as live imports.",
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": false,
              "description": "Validation-only mode: validate all rows WITHOUT persisting to database. Returns the same field/row-level errors as live imports.",
              "title": "Dry Run",
              "type": "boolean"
            }
          },
          {
            "description": "Write consistency mode for live imports. atomic=all-or-nothing, partial=persist valid rows and report row-level failures.",
            "in": "query",
            "name": "consistency",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ImportConsistency",
              "default": "atomic",
              "description": "Write consistency mode for live imports. atomic=all-or-nothing, partial=persist valid rows and report row-level failures."
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_outages_api_v1_outages_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk import outages from CSV or JSON file with optional dry-run validation and explicit consistency mode",
        "tags": [
          "outages"
        ]
      }
    },
    "/api/v1/outages/violations": {
      "get": {
        "operationId": "list_violations_api_v1_outages_violations_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Violations",
        "tags": [
          "outages"
        ]
      }
    },
...
  }
}